+++ src/pipeline/build.ts
@@ -65,6 +65,12 @@ export function enableFeatures(
             name: "alwaysSaveContractData",
         },
         {
+            option:
+                config.options.optimizations
+                    ?.internalExternalReceiversOutsideMethodsMap ?? false,
+            name: "internalExternalReceiversOutsideMethodsMap",
+        },
+        {
             option: config.options.enableLazyDeploymentCompletedGetter ?? false,
             name: "lazyDeploymentCompletedGetter",
         },